requirements.txt
asgiref== 3.6.0
django==4.1.8
django-computedfields==0.2.1
django-cors-headers==3.14.0
django-fast-update==0.2.1
django-filter==23.1
django-mysql==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
mysqlclient==2.1.1
pillow==9.5.0
pyjwt==2.6.0
pytz==2023.3
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2023.3

